# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Medical System API",
    "version": "1.0.0",
    "description": "API documentation for the Medical System"
  },
  "servers": [
    {
      "url": "http://localhost:3030"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/appointments": {
      "get": {
        "summary": "Get all appointments",
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "A list of appointments"
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "tags": [
          "Appointments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patientName",
                  "doctorId",
                  "date"
                ],
                "properties": {
                  "patientName": {
                    "type": "string"
                  },
                  "doctorId": {
                    "type": "number"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully"
          }
        }
      }
    },
    "/appointments/{id}": {
      "get": {
        "summary": "Get appointment by ID",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment found"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      },
      "delete": {
        "summary": "Delete appointment by ID",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Appointment deleted successfully"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
    "/appointments/{id}/status": {
      "put": {
        "summary": "Update appointment status",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "confirmed",
                      "cancelled",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
...
    "/complaints": {
      "post": {
        "summary": "Submit a new complaint (any logged-in user)",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message",
                  "userId"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint submitted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all complaints (admin only)",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of complaints"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/complaints/{id}": {
      "get": {
        "summary": "Get a complaint by ID (admin only)",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complaint found"
          },
          "404": {
            "description": "Complaint not found"
          }
        }
      },
      "put": {
        "summary": "Update complaint status (admin only)",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "open",
                      "resolved",
                      "rejected"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complaint status updated"
          },
          "404": {
            "description": "Complaint not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a complaint by ID (admin only)",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Complaint deleted"
          },
          "404": {
            "description": "Complaint not found"
          }
        }
      }
    },
    "/doctors": {
      "get": {
        "summary": "Get all doctors",
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "List of doctors"
          }
        }
      },
      "post": {
        "summary": "Create a new doctor (admin only)",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "specialization"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "specialization": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Doctor created"
          },
          "403": {
            "description": "Forbidden (admin only)"
          }
        }
      }
    },
    "/doctors/{id}": {
      "get": {
        "summary": "Get a doctor by ID",
        "tags": [
          "Doctors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor found"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      },
      "put": {
        "summary": "Update doctor info (admin only)",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "specialization": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Doctor updated"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a doctor (admin only)",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor deleted"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
...
  },
...
}